standard assembly register names
changed from the non-standard $n format for indexing registers to the standard names (it was a good concept but faced limitations as instructions became more diverse).
new register form is now standard-conforming and allows for a few optimizations in parsing (esp. in memory addresses, since standard register names are legal symbol names).

removed all the extend instructions (non-standard).
MOVZX and MOVSX will replace them for integers.
floating point is soon to be entirely rewritten as x87.

added BT/BTS/BTR/BTC instructions.

changed register subdivision typings to reflect their size.
added a register subdivision for the high byte of the low word (for AH, BH, CH, and DH).

renamed FlagsRegister.Flags to RFLAGS.
added FlagsRegister.EFLAGS to get the lower dword.
added FlagsRegister.WFLAGS (not a standard name) to get the lower word.

+++ src/GraphicalClient.cs
@@ -246,13 +246,13 @@ namespace CSX64
                 // -- data utilities -- //
 
                 case (UInt64)GraphicalSyscallCodes.GetRenderSize:
-                    GetRegister(1).x32 = (UInt64)RenderImage.Width;
-                    GetRegister(2).x32 = (UInt64)RenderImage.Height;
+                    GetRegister(1).x32 = (UInt32)RenderImage.Width;
+                    GetRegister(2).x32 = (UInt32)RenderImage.Height;
                     break;
 
                 case (UInt64)GraphicalSyscallCodes.GetMousePos:
-                    GetRegister(1).x32 = (UInt64)MousePos.X;
-                    GetRegister(2).x32 = (UInt64)MousePos.Y;
+                    GetRegister(1).x32 = (UInt32)MousePos.X;
+                    GetRegister(2).x32 = (UInt32)MousePos.Y;
                     break;
                 case (UInt64)GraphicalSyscallCodes.GetMouseDelta:
                     GetFlags().ZF = (GetRegister(1).x64 = (UInt64)MouseDelta) == 0;
